+++ Resources/Prototypes/Actions/polymorph.yml
@@ -43,6 +43,7 @@
   components:
   - type: Magic # DS14
   - type: Action
+    useDelay: 300 # DS14
     icon:
       sprite: Objects/Fun/immovable_rod.rsi
       state: icon
